Revert "add spinalcordtoolbox.metrics with msct_image.compute_dice moved into it" (jca working on metrics)
This reverts commit 4b2c3809de5a5d0fe6a68b01b3f2863b44112c2a [formerly 9523ff2313d749a6965ec1fcd4c603549357fcba] [formerly a5d8fa43c95c6ae7ad28e571a060d93e71ee0266] [formerly 15972b93121899526ca3f017484bb0adfca81c7b [formerly 3f7b819c6a554f7a7323c31e834670df90856617]].


Former-commit-id: d8b3399c34af2433d2f5ee32fc7173d97dd7aa3d [formerly 13b68ccd82a8ddbb66839a4628be3fce96fce859]
Former-commit-id: 93417ce8fbb30dd9f185196618ceff423adbba7e
Former-commit-id: 2f526294ba40c5cfcc8b198a9bdf056f1eb2f920
